requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.1
langchain-chroma==0.2.1
langchain-huggingface==0.1.2
chromadb==0.6.3
openai==1.59.9
sentence-transformers==3.4.1
tiktoken==0.8.0
faiss-cpu==1.9.0.post1
pypdf==5.1.0
python-dotenv==1.0.1
pydantic==2.10.4
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
